Don't duplicate import/export/func count checking in metadce tests
For imports, exports and functions we do a file comparison.  In this
case there the number of import, exports or functions is defined
by the number of lines in the file.  There is no need to duplicate
this function in the test code.  This also makes the tests easier to
update via `EMTEST_REBASELINE`.

Also add expectation files for the the libcxx tests that were present
for the fastcomp for disabled for llvm for some reason.